Implemented core, manually created scripts, misc stuff
>Bugfixes

-Fixed string variables being empty when casting from LEX::String

-Fixed api::container(const container&) using the size of other instead of it's data

-Added extra checks in variable's cast function to prevent it from returning the default value

-Fixed issue in qualified types where explicit conversions weren't being marked properly

-Fixed issue in NumberTypes where explicit conversions from float to integer weren't being respected

-Fixed bug in test run where std::cin was used rather than getline

-Fixed bug where RefreshLinkage would only fire when not yet linked
==========================

>Changes
-Changed CreateScript to allow for the content as a string, allowing for scripts to be manually loadable by string

-Increased Line's size to 32 bit and SyntaxType's size to 16 bit (Helps accommodate the appendage scripts)

-The script syntax's column and line now store the last column and line of the script.

-Added correct pages to allow appended content to scale with the current line count.

=========================================
>Additions

-Added Core, a script that's inaccessible through normal means but is able to be appended to based on the core directory. All elements in core are aggregates of the core directory
-Added IsIncremental/SetIncremental, used to tell if a script is incremental or not.

+++ include/Lexicon/TypeAliases.h
@@ -5,7 +5,7 @@
 namespace LEX
 {
 	using Column = uint16_t;
-	using Line = uint16_t;
+	using Line = uint32_t;
 
 	//Temporary
 	//using Object = int;
